+++ .bumpversion.cfg
@@ -8,7 +8,7 @@ search = version='{current_version}'
 replace = version='{new_version}'
 
 
-[bumpversion:file:emboa/version.py]
+[bumpversion:file:emperator/version.py]
 search = __version__ = '{current_version}'
 replace = __version__ = '{new_version}'
