--- .travis.yml
@@ -1,51 +1,0 @@
-language: python
-dist: xenial
-python: 3.7
-env:
-matrix:
-  fast_finish: true
-  include:
-    - os: linux
-    #- os: osx
-env:
-  global:
-    - MINCONDA_VERSION="latest"
-    - MINCONDA_LINUX="Linux-x86_64"
-    - MINCONDA_OSX="MacOSX-x86_64"
-
-install:
-    # Define the version of miniconda to download
-    - if [[ "$TRAVIS_OS_NAME" == "linux" ]]; then
-        MINCONDA_OS=$MINCONDA_LINUX;
-      elif [[ "$TRAVIS_OS_NAME" == "osx" ]]; then
-        MINCONDA_OS=$MINCONDA_OSX;
-      fi
-    - wget "http://repo.continuum.io/miniconda/Miniconda3-$MINCONDA_VERSION-$MINCONDA_OS.sh" -O miniconda.sh;
-    - bash miniconda.sh -b -p $HOME/miniconda
-    - export PATH="$HOME/miniconda/bin:$PATH"
-    - hash -r
-    - conda config --set always_yes yes --set changeps1 no
-    - conda update -q conda
-
-script:
-- chmod +x .ci/test_project.sh
-- ./.ci/test_project.sh
-
-
-# Assuming you have installed the travis-ci CLI tool, after you
-# create the Github repo and add it to Travis, run the
-# following command to finish PyPI deployment setup:
-# $ travis encrypt --add deploy.password
-# to install the travis cmd tool try
-# $ gem install travis
-# (https://docs.travis-ci.com/user/encryption-keys/)
-# deploy:
-#   provider: pypi
-#   distributions: sdist bdist_wheel
-#   user: DerThorsten
-#   password:
-#     secure: PLEASE_REPLACE_ME
-#   on:
-#     tags: true
-#     repo: DerThorsten/emboa
-#     python: 3.7
