+++ README.md
@@ -1,4 +1,4 @@
-# emboa
+# emperator
 
 Pack a mamba environment with emsdk
 
@@ -15,7 +15,7 @@ emsdk activate 3.1.2
 ## Installation
 
 ```bash
-pip install git+https://github.com/emscripten-forge/emboa
+pip install git+https://github.com/emscripten-forge/emperator
 ```
 
 ## Usage
@@ -23,7 +23,7 @@ pip install git+https://github.com/emscripten-forge/emboa
 You can pack the Python 3.10 environment (located at `/path/to/env`) with the following command:
 
 ```bash
-emboa pack python core /path/to/env --version=3.10
+emperator pack python core /path/to/env --version=3.10
 ```
 
 This will generate two files `python_data.js` and `python_data.data` that you can use in the browser.
