+++ setup.py
@@ -40,22 +40,22 @@ setup(
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
     ],
-    description="emboa emscripten+boa",
+    description="emperator emscripten+boa",
     entry_points={
         'console_scripts': [
-            'emboa=emboa.cli.main:app',
+            'emperator=emperator.cli.main:app',
         ],
     },
     install_requires=requirements,
     long_description=readme,
     include_package_data=True,
-    keywords='emboa',
-    name='emboa',
+    keywords='emperator',
+    name='emperator',
     packages=find_packages(),
     setup_requires=setup_requirements,
     test_suite='tests',
     tests_require=test_requirements,
-    url='https://github.com/DerThorsten/emboa',
+    url='https://github.com/emscripten-forge/emperator',
     version='0.1.0',
     zip_safe=False,
 )
