+++ tests/test_version.py
@@ -1,8 +1,8 @@
-import emboa
+import emperator
 
 
 class TestVersion(object):
 
     def test_version(self):
-        v = emboa.__version__
+        v = emperator.__version__
         assert v == '0.1.0'
